+++ app/src/main/java/heroiceraser/mulatschak/game/GamePlay/Background4Player0Animations.java
@@ -38,7 +38,7 @@ public class Background4Player0Animations {
     //
     public void init(GameLayout layout) {
         background.set(0, layout.getSectors().get(2).y, layout.getScreenWidth(),
-                (int) (layout.getSectors().get(7).y));
+                layout.getSectors().get(7).y);
         backgroundPaint.setColor(Color.BLACK);
         backgroundPaint.setAlpha(ALPHA_MAX);
     }
